refactor: move /me endpoint from AuthController to UserController
