+++ backend/account-service/src/main/java/com/ita/challenges/account/infrastructure/adapter/web/in/AuthController.java
@@ -60,27 +60,4 @@ public class AuthController {
                 )
         );
     }
-
-    @GetMapping("/auth/me")
-    public ResponseEntity<AuthUserDto> authMe(@AuthenticationPrincipal OAuth2User user) {
-        if (user == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-
-        return ResponseEntity.ok(mapUser(user));
-
-    }
-
-        private AuthUserDto mapUser (OAuth2User user){
-
-            String login = user.getAttribute("login");
-            String avatarUrl = user.getAttribute("avatar_url");
-
-            if (login == null) {
-                login = "unknown";
-            }
-
-            return new AuthUserDto(login, avatarUrl);
-        }
-
-    }
+}
