+++ backend/account-service/src/main/java/com/ita/challenges/account/infrastructure/adapter/web/in/UserController.java
@@ -1,8 +1,11 @@
 package com.ita.challenges.account.infrastructure.adapter.web.in;
 
 import com.ita.challenges.account.domain.port.out.UserRepository;
+import com.ita.challenges.account.infrastructure.adapter.web.in.dto.AuthUserDto;
 import com.ita.challenges.account.infrastructure.adapter.web.in.dto.UserRoleResponse;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -26,5 +29,24 @@ public class UserController {
         return userRepository.findByUsername(username)
                 .map(user -> ResponseEntity.ok(new UserRoleResponse(user.userRole())))
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
+    }
+
+    @GetMapping("/me")
+    public ResponseEntity<AuthUserDto> authMe(@AuthenticationPrincipal OAuth2User user) {
+        if (user == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        return ResponseEntity.ok(mapUser(user));
+    }
+
+    private AuthUserDto mapUser(OAuth2User user) {
+        String login = user.getAttribute("login");
+        String avatarUrl = user.getAttribute("avatar_url");
+
+        if (login == null) {
+            login = "unknown";
+        }
+
+        return new AuthUserDto(login, avatarUrl);
     }
 }
